Profile -> collection
Rename profile to collection and add some clarifications. Just a first pass to throw stones at.

Note that I retained/repurposed the word "profile" only for informal use in the example of a "professional profile", which seemed the best way to capture everyone's mental model (e.g. "LinkedIn Profile").

Was on the fence about whether to include the following elements from the discussion in #47, but decided it was _probably_ unnecessary. 

1. Additional clarifying examples:
| Number | Collection  | Collection name | Description of Collection / "Profile" | 
|---| ---|---|----|
| 1 | credentials issued by issuer A | "Issuer A" | Information about issuer A (type `Organization`) |
| 2 | Credentials related to my work history | "Kim's Verifiable Work Record" | Information about me (type `Person`) |  
| 3 | all issuers/orgs that have issued me credentials | "My Issuers" | maybe some wallet-generated metadata | 

2. Clarifying that collection contents may also be populated dynamically (by facet, tag, whatever).